+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/cucxabong/aws-google-login
+module github.com/ChatLingual/aws-google-login
 
 go 1.15
 
@@ -7,7 +7,7 @@ require (
 	github.com/aws/aws-sdk-go v1.36.14
 	github.com/kardianos/osext v0.0.0-20190222173326-2bc1f35cddc0 // indirect
 	github.com/manifoldco/promptui v0.8.0
-	github.com/mxschmitt/playwright-go v0.1400.0
 	github.com/nu7hatch/gouuid v0.0.0-20131221200532-179d4d0c4d8d // indirect
+	github.com/playwright-community/playwright-go v0.3900.1
 	github.com/urfave/cli/v2 v2.3.0
 )
