+++ google.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"net/url"
 
-	"github.com/mxschmitt/playwright-go"
+	"github.com/playwright-community/playwright-go"
 )
 
 type Google struct {
@@ -55,7 +55,14 @@ func (g *Google) Login() (string, error) {
 		return SAMLResponse, fmt.Errorf("unable to click on email field")
 	}
 
-	r := page.WaitForRequest(g.WaitURL())
+	r, err := page.ExpectRequest(g.WaitURL(), func() error {
+		return err
+	})
+
+	if err != nil {
+		return SAMLResponse, fmt.Errorf("can not ExpectRequest %v", err)
+	}
+
 	data, err := r.PostData()
 	if err != nil {
 		return SAMLResponse, fmt.Errorf("can not get PostData %v", err)
